Add support in ThresholdAnalyzer for "cross_run" regression detection.
When the ThresholdConfig's DataFilter.DataType is METRIC_SAMPLEPOINTS this will combine the current run's sample points with the sample points from all historic runs and then execute the threshold analyzer over the results.

When the ThresholdConfig's DataFilter.DataType is not METRIC_SAMPLEPOINTS it will use the median of the pre-computed aggregates for the specified DataFilter in each run from the set of historical runs and the current run being analyzed.

Change-Id: I5d50f031762e2d9248f21a6b12cfea60ff4acff6