Fix issue with medians being incorrectly computed in CrossRunThreshold functionality of ThresholdAnalyzer.
The issue is that we were sorting the collection of data points on the first element of a pair, which is the timestamp field (the second element of pair is the metric value). Fix the sort function to sort on the second element in the pair so that the middle element int he sorted collection is the median metric value.

PiperOrigin-RevId: 279358763
Change-Id: I70f1b93b20146282f5d6a18295f63c50f804dec9